pyproject.toml
[tool.poetry]
name = "azure-emulators-tests"
version = "0.1.0"
description = "Tests for Azure emulators in Docker"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
azure-cosmos = "^4.3.0"
azure-storage-blob = "^12.13.0"
azure-storage-queue = "^12.6.0"
azure-storage-table = "^12.0.0"
azure-servicebus = "^7.8.0"
pytest = "^7.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"